Fix sign comparison warnings.
git-svn-id: https://geany.svn.sourceforge.net/svnroot/geany/trunk@947 ea778897-0a13-0410-b9d1-a72fbfd435f5

+++ src/encodings.c
@@ -346,7 +346,7 @@ gchar *encodings_convert_to_utf8(const gchar *buffer, gsize size, gchar **used_e
 
 	for (i = 0; i < GEANY_ENCODINGS_MAX; i++)
 	{
-		if (i == encodings[GEANY_ENCODING_NONE].idx) continue;
+		if (i == (guint) encodings[GEANY_ENCODING_NONE].idx) continue;
 		
 		if (check_current)
 		{
